% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{mj_attributes}
\alias{mj_attributes}
\alias{mj_all}
\alias{mj_class}
\title{Modify defaults}
\usage{
mj_attributes(...)

mj_all(...)

mj_class(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
This tag allows you to modify default attributes on a \code{ml_tag} and add \code{mj_class} to them.
}
\section{Functions}{

\itemize{
  \item{\code{mj_attributes} }{Wrap all your attributes.}
  \item{\code{mj_class} }{Add a class.}
  \item{\code{mj_all} }{Set default attributes for every components inside your email.}
}
}

\examples{
mj_ml(
  mj_head(
    mj_attributes(
      mj_text(padding = 0),
      mj_class(name = "blue", color = "blue"),
      mj_class(name = "big", `font-size` = "20px"),
      mj_all(`font-family` = "Arial")
    )
  ),
  mj_body(
    mj_container(
      mj_section(
        mj_column(
          mj_text(
            `mj-class` = "blue big",
            "Blue text in Arial and 20px"
          )
        )
      )
    )
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-attributes}
}
