% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{mj_font}
\alias{mj_font}
\alias{mj_font_google}
\title{Font}
\usage{
mj_font(href, name)

mj_font_google(name)
}
\arguments{
\item{href}{URL of the font.}

\item{name}{Name of the font.}
}
\description{
This tag allows you to import fonts if used in your MJML document.
}
\details{
Use \code{mj_font_google} for \href{Google Fonts}{https://fonts.google.com/}.
}
\examples{
mj_ml(
 mj_head(
   mj_font(
     href = "https://fonts.googleapis.com/css?family=Raleway",
     name = "Raleway"
   )
 ),
 mj_body(
    mj_text(
      `font-family` = "Raleway, Arial",
      "Blue text"
    )
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-font},
\href{Google Fonts}{https://fonts.google.com/}
}
