% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hero.R
\name{mj_hero}
\alias{mj_hero}
\alias{mj_hero_content}
\title{Hero}
\usage{
mj_hero(...)

mj_hero_content(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Display a section with a background image and some content inside (\code{\link{mj_text}},
\code{\link{mj_button}}, \code{\link{mj_image}} ...) wrapped in mj-hero-content component.
}
\examples{
mj_ml(
  mj_body(
    mj_container(
      mj_hero(
        mode = "fluid-height",
        `background-url` = paste0("https://media.vanityfair.com/photos/",
                                  "54cbf3da998d4de83ba3602a/master/",
                                  "w_960,c_limit/image.jpg"),
        mj_hero_content(
          mj_text("HERO")
        ),
        mj_button("MJML")
      )
    )
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-hero}
}
