% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoice.R
\name{mj_invoice}
\alias{mj_invoice}
\alias{mj_invoice_item}
\title{Invoice}
\usage{
mj_invoice(...)

mj_invoice_item(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Display a table of items with calculated total price.
}
\examples{
mj_ml(
  mj_body(
    mj_container(
      mj_section(
        mj_column(
          mj_invoice(
            format = "0,00.00€",
            intl = "name:Product Name",
            mj_invoice_item(name = "TV", price = "549€", quantity = "1"),
            mj_invoice_item(name = "DVD - Iron Man II", price = "22.99€", quantity = "2")
          )
        )
      )
    )
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-invoice}
}
