% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{mj_button}
\alias{mj_button}
\title{Button}
\usage{
mj_button(href = "#", ...)
}
\arguments{
\item{href}{href link.}

\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Add a button.
}
\examples{
mj_ml(
  mj_body(
    mj_container(
      mj_button("http://john-coene.com", "Button")
    )
  )
)

}
