% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{mj_chart}
\alias{mj_chart}
\title{Chart}
\usage{
mj_chart(cht = "bvg", ...)
}
\arguments{
\item{cht}{chart type.}

\item{...}{any other option, see \href{official documentation}{https://github.com/image-charts/mjml-chart}.}
}
\description{
Chart
}
\examples{
mj_ml(
  mj_body(
    mj_container(
      mj_section(
        mj_column(
          mj_chart(
            chd = "t:10,20,30|15,25,35"
          )
        )
      )
    )
  )
)

}
\seealso{
\href{official documentation}{https://github.com/image-charts/mjml-chart}
}
