% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{mj_column}
\alias{mj_column}
\title{Layout}
\usage{
mj_column(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Create colmns, sections and containers.
}
\section{Function}{

\itemize{
  \item{\code{mj_section} }{Horizontal sections.}
  \item{\code{mj_container} }{Standard container.}
  \item{\code{mj_column} }{Column, up to four.}
}
}

\examples{
mj_ml(
  mj_body(
    mj_container(
      mj_section(
        mj_column(
          mj_text("First Column")
        ),
        mj_column(
          mj_text("Second Column")
        )
      ),
      mj_section(
        mj_column(
          mj_text("First Column")
        ),
        mj_column(
          mj_text("Second Column")
        ),
        mj_column(
          mj_text("Third Column")
        )
      )
    )
  )
)

}
