% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{mj_container}
\alias{mj_container}
\alias{mj_container}
\title{Create a container}
\usage{
mj_container(...)

mj_container(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Create a container.

Create a section.
}
