% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{mj_divider}
\alias{mj_divider}
\title{Divider}
\usage{
mj_divider(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Add a divider.
}
\examples{
mj_ml(
  mj_head(
    mj_text("The header")
  ),
  mj_body(
    mj_divider(),
    mj_text("The body")
  )
)

}
