% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{mj_font}
\alias{mj_font}
\title{Font}
\usage{
mj_font(href, name)
}
\arguments{
\item{href}{URL of the font.}

\item{name}{Name of the font.}
}
\description{
This tag allows you to import fonts if used in your MJML document.
}
\examples{
mj_ml(
 mj_head(
   mj_font(
     href = "https://fonts.googleapis.com/css?family=Raleway",
     name = "Raleway"
   )
 ),
 mj_body(
    mj_text(
      `font-family` = "Raleway, Arial",
      "Blue text"
    )
  )
)

}
