% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{mj_group}
\alias{mj_group}
\title{Group}
\usage{
mj_group(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Allows you to prevent columns from stacking on mobile. To do so, wrap the columns inside
a mj-group tag, so they'll stay side by side on mobile.
}
\examples{
mj_ml(
  mj_body(
    mj_container(
      mj_group(
        mj_column(
          mj_text("Some text")
        ),
        mj_column(
          mj_text("Some more text")
        )
      )
    )
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-group}
}
