% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{mj_image}
\alias{mj_image}
\title{Image}
\usage{
mj_image(src, ...)
}
\arguments{
\item{src}{image source.}

\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Displays a responsive image in your email. It is similar to the HTML
\code{<img/>} tag. Note that if no width is provided, the image will use the parent column width.
}
\examples{
mj_ml(
  mj_body(
    mj_image(paste0("https://media.vanityfair.com/photos/",
                    "54cbf3da998d4de83ba3602a/master/w_960,c_limit/image.jpg"))
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-image}
}
