% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{mj_location}
\alias{mj_location}
\title{Location}
\usage{
mj_location(address, ...)
}
\arguments{
\item{address}{address.}

\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Display a Google Maps location link.
}
\examples{
mj_ml(
  mj_body(
    mj_column(
      mj_location("Times Square")
    )
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-location}
}
