% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{mj_preview}
\alias{mj_preview}
\title{Preview}
\usage{
mj_preview(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
This tag allows you to set the preview that will be displayed in the inbox of the recipient.
}
\examples{
mj_ml(
  mj_head(
    mj_preview("Hello MJML")
  )
)

}
