% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{mj_raw}
\alias{mj_raw}
\title{Raw}
\usage{
mj_raw(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Displays raw HTML that is not going to be parsed by the MJML engine.
Anything left inside this tag should be raw, responsive HTML.
}
\examples{
mj_ml(
  mj_body(
    mj_column(
      mj_raw("<h1>Raw HTML</h1>")
    )
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-raw}
}
