% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjml.R
\name{mj_save}
\alias{mj_save}
\title{Save}
\usage{
mj_save(mjml, file, ...)
}
\arguments{
\item{mjml}{MJML email as returned by \code{\link{mj_ml}}.}

\item{file}{File name of output.}

\item{...}{Any other option to be passed to \link[htmltools]{save_html}.}
}
\description{
Save your email.
}
