% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{mj_section}
\alias{mj_section}
\title{Section}
\usage{
mj_section(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Sections are intended to be used as rows within your email. They will be used to structure the layout.
}
\examples{
mj_ml(
  mj_body(
    mj_container(
      mj_section(
        mj_text("Body")
      )
    )
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-section}
}
