% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{mj_social}
\alias{mj_social}
\title{Social}
\usage{
mj_social(display, ...)
}
\arguments{
\item{display}{icon to display.}

\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Displays calls-to-action for various social networks with their associated logo.
You can activate/deactivate any icon, with \code{display} property.
}
\examples{
mj_ml(
  mj_body(
    mj_social("facebook")
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-social}
}
