% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{mj_spacer}
\alias{mj_spacer}
\title{Spacer}
\usage{
mj_spacer(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Displays a blank space.
}
\examples{
mj_ml(
  mj_body(
    mj_divider(),
    mj_text("Some text"),
    mj_spacer(),
    mj_text("Some more text")
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-spacer}
}
