% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{mj_style}
\alias{mj_style}
\title{Style}
\usage{
mj_style(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
#' Style
}
\details{
This tag allows you to set CSS styles that will be applied to the HTML in your MJML document
as well as the HTML outputted. The CSS styles will be added to the head of the rendered HTML
by default, but can also be inlined by using the \code{inline="inline"} attribute.
}
\examples{
mj_ml(
 mj_head(
   mj_style(
     inline = TRUE,
     ".color{color:'blue'}"
   )
 ),
 mj_body(
    mj_text(
      `mj-class` = "color",
      "Blue text"
    )
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-style}
}
