% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{mj_title}
\alias{mj_title}
\title{Title}
\usage{
mj_title(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
This tag allows you to set the title of an MJML document.
}
\examples{
mj_ml(
  mj_head(
    mj_title("Hello MJML")
  )
)

}
