% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{mj_view}
\alias{mj_view}
\title{Viewer}
\usage{
mj_view(mjml)
}
\arguments{
\item{mjml}{An \code{mjml} object as returned by \code{\link{mj_ml}.}}
}
\description{
Preview email in the RStudio viewer
}
\examples{
if(interactive()){
mj_ml(
  mj_body(
    mj_column(
      width = "20\%",
      mj_text("First columns")
    ),
    mj_column(
      width = "80\%",
      mj_text("The body")
    )
  )
) \%>\%
  mj_view()
}

}
