% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{mj_wrapper}
\alias{mj_wrapper}
\title{Wrapper}
\usage{
mj_wrapper(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Wrapper enables to wrap multiple sections together. It's especially useful to achieve nested
layouts with shared border or background images across sections.
}
\examples{
mj_ml(
  mj_body(
    mj_container(
      mj_wrapper(
        mj_section(
          mj_column(
            mj_text("Some text")
          )
        ),
        mj_section(
          mj_column(
            mj_text("Text"),
            mj_divider(),
            mj_text("More text")
          )
        )
      )
    )
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-wrapper}
}
