% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mjml}
\alias{mjml}
\alias{mjml-package}
\title{mjml: A package for building and sending responsive emails.}
\description{
The package provides an API to all \href{mjml}{https://mjml.io} tags.
}
\section{Includes templates as RStudio addins.}{
NA
}

