% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{mj_navbar}
\alias{mj_navbar}
\alias{mj_link}
\alias{mj_inline_links}
\title{Navbar}
\usage{
mj_navbar(...)

mj_link(...)

mj_inline_links(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Displays a menu for navigation with an optional hamburger mode for mobile devices.
}
\examples{
mj_ml(
  mj_body(
    mj_navbar(
      mj_column(
        width = "20\%",
        mj_image(
          paste0("https://media.vanityfair.com/photos/",
          "54cbf3dafde9250a6c402ed2/master/w_690,c_limit/image.jpg")
        )
      ),
      mj_column(
        width = "80\%",
        mj_inline_links(
          `base-url` = "http://john-coene.com",
          mj_link(
            href = "/datamaps"
          ),
          mj_link(
            href = "/packages/fullPage/"
          )
        )
      )
    )
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-navbar}
}
