% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{mj_send_html}
\alias{mj_send_html}
\alias{mj_send}
\title{Send}
\usage{
mj_send_html(email, from, to, subject, html = TRUE, inline = TRUE,
  authenticate = TRUE, send = TRUE, ...)

mj_send(mjml, from, to, subject, html = TRUE, inline = TRUE,
  authenticate = TRUE, send = TRUE, ...)
}
\arguments{
\item{email, mjml}{Email full path or \code{\link{mj_ml}} object.}

\item{from}{From whom the mail message is (RFC2822 style address).}

\item{to}{Recipient of the message (vector of valid RFC2822 style addresses).}

\item{subject}{Subject line of message.}

\item{html}{A boolean indicating whether the body of the email should be parsed as HTML.}

\item{inline}{A boolean indicating whether images in the HTML file should be embedded inline.}

\item{authenticate}{A boolean variable to indicate whether authorization is required to connect to the SMTP server.
If set to \code{TRUE}, see details on parameters required in smtp parameter.}

\item{send}{A boolean indicating whether the email should be sent at the end of the function (default behaviour).
If set to \code{FALSE}, function returns the email object to the parent environment.}

\item{...}{any other option to pass to \link[mailR]{send.mail}.}
}
\description{
Send email via \link[mailR]{send.mail}.
}
\section{Functions}{

\itemize{
  \item{\code{mj_send_html} }{Send email from \code{.html} file as returned by \code{\link{mj_convert_html}}}
  \item{\code{mj_send} }{Send email}
}
}

\examples{
\dontrun{
mj_ml(
  mj_body(
    mj_container(
      mj_section(
        mj_column(
          mj_text("First column")
        ),
        mj_column(
          mj_text("Second column")
        )
      ),
      mj_section(
        mj_column(
          mj_image(paste0("https://media.vanityfair.com/photos/",
                          "54cbf3da998d4de83ba3602a/master/w_960,",
                          "c_limit/image.jpg")
                   )
        )
      )
    )
  )
) \%>\%
  mj_save("email") \%>\%
  mj_send_html("me@gmail.com", "him@gmail.com", "MJML email",
               authenticate = TRUE,
               send= TRUE,
               smtp = list(host.name = "smtp.gmail.com",
                           port = 465,
                           user.name = "gmail_username",
                           passwd = "password",
                           ssl = TRUE)
               )
}

}
