% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjml.R
\name{mj_set}
\alias{mj_set}
\alias{mj_get}
\alias{install_mjml}
\title{Setup}
\usage{
mj_set(path)

mj_get()

install_mjml()
}
\arguments{
\item{path}{Path to MJML.}
}
\description{
Set path to MJML.
}
\examples{
\dontrun{
install_mjml()
}

}
