% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{mj_column}
\alias{mj_column}
\title{Column}
\usage{
mj_column(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Columns enable you to horizontally organize the content within your sections.
They must be located under mj-section tags in order to be considered by the engine.
To be responsive, columns are expressed in terms of percentage. Every single column has
to contain something because they are responsive containers, and will be vertically stacked
on a mobile view. Any standard component, or component that you have defined and registered,
can be placed within a column – except mj-column or mj-section elements.
}
\section{Function}{

\itemize{
  \item{\code{mj_section} }{Horizontal sections.}
  \item{\code{mj_container} }{Standard container.}
  \item{\code{mj_column} }{Column, up to four.}
}
}

\examples{
mj_ml(
  mj_body(
    mj_container(
      mj_section(
        mj_column(
          mj_text("First Column")
        ),
        mj_column(
          mj_text("Second Column")
        )
      ),
      mj_section(
        mj_column(
          mj_text("First Column")
        ),
        mj_column(
          mj_text("Second Column")
        ),
        mj_column(
          mj_text("Third Column")
        )
      )
    )
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-column}
}
