% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{mj_column}
\alias{mj_column}
\title{Create columns}
\usage{
mj_column(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Create up to 4 columns.
}
\examples{
mj_ml(
  mj_body("")
)

}
