% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{mj_container}
\alias{mj_container}
\title{Container}
\usage{
mj_container(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Create a container.
}
\examples{
mj_ml(
  mj_body(
    mj_container(
      mj_text("Some text")
    )
  )
)

}
