% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{mj_divider}
\alias{mj_divider}
\title{Divider}
\usage{
mj_divider(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Displays a horizontal divider that can be customized like a HTML border.
}
\examples{
mj_ml(
  mj_body(
    mj_text("Some text"),
    mj_divider(),
    mj_text("The rest of the body")
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-divider}
}
