% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjml.R
\name{mj_save}
\alias{mj_save}
\title{Save}
\usage{
mj_save(mjml, file, ...)
}
\arguments{
\item{mjml}{MJML email as returned by \code{\link{mj_ml}}.}

\item{file}{File name of output.}

\item{...}{Any other option to be passed to \code{\link{write}}.}
}
\description{
Save your email with \code{.mjml} extension.
}
\examples{
mj_ml(
  mj_body(
    mj_container(
      mj_section(
        mj_column("First column"),
        mj_column("Second column")
      ),
      mj_section(
        mj_chart(
          chd = "t:10,20,30|15,25,35"
        )
      )
    )
  )
) \%>\%
  mj_save("email.mjml")

}
