% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{mj_social}
\alias{mj_social}
\title{Add icons}
\usage{
mj_social(display, ...)
}
\arguments{
\item{display}{icon to display.}

\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Add icons.
}
\examples{
mj_ml(
  mj_body(
    mj_social("facebook")
  )
)

}
