% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{mj_social}
\alias{mj_social}
\title{Social}
\usage{
mj_social(mode = "horizontal", ...)
}
\arguments{
\item{mode}{Display mode.}

\item{...}{Unnamed items that comprise this list of MJML tags.}

\item{name}{Name of social element, e.g.: \code{facebook}.}
}
\description{
Displays calls-to-action for various social networks with their associated logo.
You can activate/deactivate any icon, with \code{display} property.
}
\examples{
mj_ml(
  mj_body(
    mj_social("facebook")
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-social}
}
