% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{mj_table}
\alias{mj_table}
\title{Table}
\usage{
mj_table(df)
}
\arguments{
\item{df}{a \code{data.frame}.}
}
\description{
This tag allows you to display table and filled it with data.
}
\examples{
mj_ml(
  mj_body(
    mj_column(
      mj_table(mtcars)
    )
  )
)

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#mjml-table}
}
