% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{mj_text}
\alias{mj_text}
\title{Text}
\usage{
mj_text(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Add text.
}
\examples{
mj_ml(
  mj_head(
    mj_attributes(
      mj_text(color = "blue")
    )
  ),
  mj_body(
    mj_container(
      mj_section(
        mj_text("Some blue text")
      )
    )
  )
)

}
