% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjml.R
\name{mj_save}
\alias{mj_save}
\alias{mj_validate}
\alias{mj_convert}
\title{Save}
\usage{
mj_save(mjml, output, ...)

mj_validate(input)

mj_convert(input, output)
}
\arguments{
\item{mjml}{MJML email as returned by \code{\link{mj_ml}}.}

\item{...}{Any other option to be passed to \code{\link{writeLines}}.}

\item{input, output}{input and output files.}
}
\description{
Save your email with \code{.mjml} extension.
}
\section{Functions}{

\itemize{
  \item{\code{mj_save} }{Save as \code{.mjml}}
  \item{\code{mj_validate} }{Validate \code{.mjml} file}
  \item{\code{mj_convert} }{Convert \code{.mjml} to \code{.html}}
}
}

\examples{
mj_ml(
  mj_body(
    mj_container(
      mj_section(
        mj_column("First column"),
        mj_column("Second column")
      ),
      mj_section(
        mj_chart(
          chd = "t:10,20,30|15,25,35"
        )
      )
    )
  )
) \%>\%
  mj_save("email.mjml") \%>\% # save as .mjml
  mj_validate() \%>\% # validate
  mj_convert("email.html") # save as .html

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#command-line-interface}
}
