% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjml.R
\name{save}
\alias{save}
\alias{mj_convert}
\alias{mj_validate_mjml}
\title{Save}
\usage{
mj_convert(mjml, output = NULL, path = NULL)

mj_validate_mjml(input, path = NULL)
}
\arguments{
\item{mjml}{MJML email as returned by \code{\link{mj_ml}}.}

\item{path}{Path to mjml installation, if \code{NULL} attempts to find
the path with \code{\link{find_mjml}}.}

\item{input, output}{input and output files.}

\item{...}{Any other option to be passed to \code{\link{writeLines}}.}
}
\value{
Path to \code{output}.
}
\description{
Save your email with \code{.mjml} extension.
}
\section{Functions}{

\itemize{
  \item{\code{mj_convert} }{Convert to \code{.html}}
  \item{\code{mj_validate_mjml} }{Validate \code{.mjml} file}
}
}

\examples{
email <- mj_ml(
  mj_body(
    mj_container(
      mj_section(
        mj_column(
          mj_text("First column")
        ),
        mj_column(
          mj_text("Second column")
        )
      ),
      mj_section(
        mj_column(
          mj_image(paste0("https://media.vanityfair.com/photos/",
                          "54cbf3da998d4de83ba3602a/master/w_960,",
                          "c_limit/image.jpg")
                   )
        )
      )
    )
  )
) 

\dontrun{html <- mj_convert(email)}

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#command-line-interface}
}
