% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjml.R
\name{mj_save_mjml}
\alias{mj_save_mjml}
\alias{mj_validate_mjml}
\alias{mj_convert_html}
\alias{mj_save}
\title{Save}
\usage{
mj_save_mjml(mjml, output, ...)

mj_validate_mjml(input)

mj_convert_html(input, output)

mj_save(mjml, output, ...)
}
\arguments{
\item{mjml}{MJML email as returned by \code{\link{mj_ml}}.}

\item{...}{Any other option to be passed to \code{\link{writeLines}}.}

\item{input, output}{input and output files.}
}
\value{
Path to \code{output}.
}
\description{
Save your email with \code{.mjml} extension.
}
\section{Functions}{

\itemize{
  \item{\code{mj_save_mjml} }{Save as \code{.mjml}}
  \item{\code{mj_validate_mjml} }{Validate \code{.mjml} file}
  \item{\code{mj_convert_html} }{Convert \code{.mjml} file to \code{.html}}
  \item{\code{mj_save} }{All of the above, outputs \code{.html}}
}
}

\examples{
mj_ml(
  mj_body(
    mj_container(
      mj_section(
        mj_column(
          mj_text("First column")
        ),
        mj_column(
          mj_text("Second column")
        )
      ),
      mj_section(
        mj_column(
          mj_image(paste0("https://media.vanityfair.com/photos/",
                          "54cbf3da998d4de83ba3602a/master/w_960,",
                          "c_limit/image.jpg")
                   )
        )
      )
    )
  )
) \%>\%
  mj_save("email")

}
\seealso{
\href{official documentation}{https://mjml.io/documentation/#command-line-interface}
}
