% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjml.R
\name{mj_set}
\alias{mj_set}
\alias{mj_get}
\title{Setup}
\usage{
mj_set(path)

mj_get()
}
\arguments{
\item{path}{Path to MJML.}
}
\description{
Set path to MJML.
}
\examples{
mj_set("./node_modules/.bin")
mj_get()

}
