% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjml.R
\name{mj_ml}
\alias{mj_ml}
\alias{mj_head}
\alias{mj_body}
\alias{mj_include}
\title{MJML}
\usage{
mj_ml(...)

mj_head(...)

mj_body(...)

mj_include(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of MJML tags.}
}
\description{
Core components of an MJML email.
}
\examples{
mj_ml(
  mj_head("header"),
  mj_body("footer")
)

}
