% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{mj_template_allura}
\alias{mj_template_allura}
\alias{mj_template_hny}
\alias{mj_template_card}
\alias{mj_template_proof}
\title{Template}
\usage{
mj_template_allura()

mj_template_hny()

mj_template_card()

mj_template_proof()
}
\description{
Responsive email templates.
}
\section{Templates}{

\itemize{
  \item{\code{mj_template_allura} {\href{Allura template}{https://mjml.io/try-it-live/templates/black-friday}}}
  \item{\code{mj_template_hny} {\href{Happy New Year template}{https://mjml.io/try-it-live/templates/happy-new-year}}}
  \item{\code{mj_template_card} {\href{Happy New Year template}{https://mjml.io/try-it-live/templates/card}}}
  \item{\code{mj_template_proof} {\href{Happy New Year template}{https://mjml.io/try-it-live/templates/proof}}}
}
}

